refactor(Function): Function 이넘 클래스 컨트롤러로 이동
- [x] 컨트롤러와 뷰에서 사용되기 때문에 controller 패키지에 둡니다.
