+++ src/main/java/janggi/controller/JanggiFunction.java
@@ -1,0 +1,7 @@
+package janggi.controller;
+
+public enum JanggiFunction {
+    MOVE,
+    GIVE_UP,
+    STOP;
+}
