--- src/main/java/janggi/domain/Function.java
@@ -1,7 +1,0 @@
-package janggi.domain;
-
-public enum Function {
-    MOVE,
-    GIVE_UP,
-    STOP;
-}
